feat(fix/auth-guard): ensure that /promotions pages handle roles correctly

+++ frontend/ZeroWaste/src/app/pages/products/list-product-page/list-product-page.component.ts
@@ -1,7 +1,7 @@
 import { CommonModule } from '@angular/common';
-import { Component, OnInit, inject, signal } from '@angular/core';
+import { Component, OnInit, ViewChild, inject, signal } from '@angular/core';
 import { ButtonComponent } from '../../../components/form/button/button.component';
-import { RouterModule } from '@angular/router';
+import { ActivatedRoute, Router, RouterModule } from '@angular/router';
 import { Product } from '../product';
 import { ProductService } from '../../../services/ProductService';
 import { ModalComponent } from "../../../components/modal/modal.component";
@@ -31,10 +31,14 @@ export class ListProductPageComponent implements OnInit {
   public productsService: ProductService = inject(ProductService);
   public fb = inject(FormBuilder);
   private validationErrorMessage = inject(ValidationErrorMessage);
+  private router = inject(Router);
+  public route = inject(ActivatedRoute);
   public products = signal<Product[]>([]);
   public filters = this.fb.group({
     daysToExpire: [null, Validators.min(0)],
   });
+
+  @ViewChild(ModalComponent) modal!: ModalComponent;
 
   public async onSubmitFilterForm(event: SubmitEvent) {
     event.preventDefault();
@@ -60,6 +64,14 @@ export class ListProductPageComponent implements OnInit {
   }
 
   public async onDeleteProductConfirmation(productId: number): Promise<void> {
+    if (this.route.snapshot.data['role'] !== 'ADMIN') {
+      alert('Você não tem permissão para deletar produtos');
+
+      this.modal.closeModal();
+
+      return;
+    }
+
     try {
       await this.productsService.deleteProduct(productId);
 
@@ -80,5 +92,4 @@ export class ListProductPageComponent implements OnInit {
 
     return validationErrorMessage;
   }
-
 }
